\name{generateCustomLib}
\alias{generateCustomLib}
\title{For a given scheme, generate a dataset with the library information}
\usage{
  generateCustomLib(Custom)
}
\arguments{
  \item{Custom}{function representing the scheme to use}
}
\value{
  A data frame of library information
}
\description{
  For a given scheme, generate a dataset with the library
  information
}

