% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateCustomNei}
\alias{generateCustomNei}
\title{For a given scheme, generate a dataset with the neighborhood information}
\usage{
generateCustomNei(scheme_def, k = 6:10, n = 6:14)
}
\arguments{
\item{scheme_def}{definition of the custom scheme}

\item{k}{peptide lengths to include}

\item{n}{exponents of the library size to include}
}
\value{
A data frame of neighborhood information
}
\description{
For a given scheme, generate a dataset with the neighborhood information
}

