\name{generateCustomProbs}
\alias{generateCustomProbs}
\title{For a given scheme, generate a dataset with the peptide probabilities}
\usage{
  generateCustomProbs(Custom)
}
\arguments{
  \item{Custom}{function representing the scheme to use}
}
\value{
  A data frame of peptide probabilities
}
\description{
  For a given scheme, generate a dataset with the peptide
  probabilities
}

