\name{getCounts}
\alias{getCounts}
\title{Get the counts possible for each scheme and k}
\usage{
  getCounts(libscheme, k)
}
\arguments{
  \item{libscheme}{The scheme to use}

  \item{k}{Peptide length}
}
\value{
  Character vector of possible counts for each class
}
\description{
  Get the counts possible for each scheme and k
}

