\docType{data}
\name{nnb_scheme}
\alias{nnb_scheme}
\title{Nucleotide library scheme NNB}
\format{'data.frame':	6 obs. of  3 variables:
 $ class: Factor w/ 6 levels "A","B","C","D",..: 1 2 3 4 5 6
 $ aacid: Factor w/ 6 levels "*C","AGPTV","DFHINY",..: 6 5 2 3 4 1
 $ c    : num  5 4 3 2 1 1.5}
\description{
  This data set contains descriptions of amino acid classes
  under the NNB library scheme.
}
\details{
  Nucleotide library scheme NNB

  The last DNA nucleus in the sequence is restricted to be
  one of C, G, or T.
}
\keyword{datasets}

