\name{NNKC}
\alias{NNKC}
\title{k-nucleotide library of scheme NNKC}
\usage{
  NNKC(k)
}
\arguments{
  \item{k}{length of peptide sequences}
}
\value{
  list consisting of a data frame of peptide classes, size
  of class, and its probabilities, and a list of additional
  information relating to the library scheme
}
\description{
  k-nucleotide library of scheme NNKC
}
\examples{
NNKC(2)
}

