\docType{data}
\name{nnnc_scheme}
\alias{nnnc_scheme}
\title{Nucleotide library scheme NNNC}
\format{'data.frame':	6 obs. of  3 variables:
 $ class : Factor w/ 6 levels "A","B","C","D",..: 1 2 3 4 5 6
 $ aacids: Factor w/ 6 levels "*","AGPTV","DEFHKNQYC",..: 6 2 4 3 5 1
 $ c     : num  6 4 3 2 1 3}
\description{
  This data set contains descriptions of amino acid classes
  under the NNN library scheme with C.
}
\details{
  Nucleotide library scheme NNNC
}
\keyword{datasets}

