\docType{data}
\name{nns_scheme}
\alias{nns_scheme}
\title{Nucleotide library scheme NNS}
\format{'data.frame':	4 obs. of  3 variables:
 $ class: Factor w/ 4 levels "A","B","C","Z": 1 2 3 4
 $ aacid: Factor w/ 4 levels "AGPTV","C*","DEFHIKMNQWY",..: 4 1 3 2
 $ c    : num  3 2 1 1}
\description{
  This data set contains descriptions of amino acid classes
  under the NNS library scheme.
}
\details{
  Nucleotide library scheme NNS

  The last DNA nucleus in the sequence is restricted to be
  one of C, or G.
}
\keyword{datasets}

