% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scheme}
\alias{scheme}
\title{Get the specified library scheme definition}
\usage{
scheme(name, file = NULL)
}
\arguments{
\item{name}{name of the scheme as a character vector}

\item{file}{CSV file hosting scheme definition, if provided}
}
\value{
a data frame of peptide classes, amino acids, and size of the classes corresponding to the selected scheme
}
\description{
Get the specified library scheme definition
}
\examples{
scheme("NNN")
scheme("NNK")
}

