\docType{data}
\name{schemes}
\alias{schemes}
\title{Built-in library schemes for peptider}
\description{
  This data set contains descriptions of amino acid classes
  several commonly used library schemes: NNN, NNB, NNK,
  trimer, and variations of each in which Cysteine is
  considered a viable amino acid.
}
\details{
  Built-in library schemes
}
\keyword{datasets}

