\docType{data}
\name{trimer_scheme}
\alias{trimer_scheme}
\title{Nucleotide library scheme Trimer}
\format{'data.frame':	2 obs. of  5 variables:
 $ class : Factor w/ 2 levels "A","Z": 1 2
 $ aacids: Factor w/ 2 levels "*","SLRAGPTVIDEFHKNQYMW": 2 1
 $ s     : num  19 0
 $ c     : num  1 0
 $ stops : num  0 0}
\description{
  This data set contains descriptions of amino acid classes
  under the Trimer library scheme.
}
\details{
  Nucleotide library scheme Trimer

  Only valid peptide sequences are created. Each sequence
  appears with the same probability.
}
\keyword{datasets}

