\docType{data}
\name{trimerc_scheme}
\alias{trimerc_scheme}
\title{Nucleotide library scheme TrimerC}
\format{'data.frame':	2 obs. of  3 variables:
 $ class : Factor w/ 2 levels "A","Z": 1 2
 $ aacids: Factor w/ 2 levels "*","SLRAGPTVIDEFHKNQYMWC": 2 1
 $ c     : num  1 0}
\description{
  This data set contains descriptions of amino acid classes
  under the TrimerC library scheme.
}
\details{
  Nucleotide library scheme Trimer with Cysteine

  Only valid peptide sequences are created. Each sequence
  appears with the same probability.
}
\keyword{datasets}

