% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{BLOSUM80}
\alias{BLOSUM80}
\title{BLOSUM80 matrix}
\usage{
data(BLOSUM80)
}
\description{
The BLOSUM80 matrix, which stands for Blocks Substitution Matrix, defines log-odds scores for the ratio of the chance of two amino acids appearing in a sequence over the chance that the two amino acids appear in any sequence.  Larger scores indicate a higher probability of substitutions.  This matrix is used in order to compute sequences which are in the neighborhood of other sequences.
}
\details{
BLOSUM80 matrix
}

