% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/peptider.r
\name{coverage}
\alias{coverage}
\title{Coverage as expected number of peptides given all possible peptides}
\usage{
coverage(k, libscheme, N, lib = NULL, variance = FALSE)
}
\arguments{
\item{k}{length of peptide sequences}

\item{libscheme}{Name (character vector) or definition (data frame) of scheme}

\item{N}{size of the library}

\item{lib}{library scheme}

\item{variance}{return the variance instead of the expected value}
}
\value{
coverage index between 0 and 1
}
\description{
Coverage of library of size N given random sampling from the pool of all possible peptides according to probabilities determined according to the library scheme.
}
\examples{
coverage(2, "NNN", 10^3)
coverage(2, "NNK", 10^3)
coverage(2, "2020", 10^3) ## 20/20 coverage is not 1 because of random sampling.
}

