% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tabulate.r
\name{encodingReduce}
\alias{encodingReduce}
\title{Reduce the regular encoding to an easier/faster format}
\usage{
encodingReduce(class, libscheme)
}
\arguments{
\item{class}{The peptide class}

\item{libscheme}{The scheme to use}
}
\value{
Vector of reduced peptide encodings
}
\description{
Reduce the regular encoding to an easier/faster format
}

