\name{generateCustomLib}
\alias{generateCustomLib}
\title{For a given scheme, generate a dataset with the library information}
\usage{
  generateCustomLib(scheme_def)
}
\arguments{
  \item{scheme_def}{definition of the custom scheme}
}
\value{
  A data frame of library information
}
\description{
  For a given scheme, generate a dataset with the library
  information
}

