% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tabulate.r
\name{generateCustomLib}
\alias{generateCustomLib}
\title{For a given scheme, generate a dataset with the library information}
\usage{
generateCustomLib(scheme_def, k = 1:20, n = 1:25)
}
\arguments{
\item{scheme_def}{definition of the custom scheme}

\item{k}{peptide lengths to include}

\item{n}{exponents of the library size to include}
}
\value{
A data frame of library information
}
\description{
For a given scheme, generate a dataset with the library information
}

