\name{generateCustomLib}
\alias{generateCustomLib}
\title{For a given scheme, generate a dataset with the library information}
\usage{
generateCustomLib(scheme_def, k = 6:10, n = 6:14)
}
\arguments{
  \item{scheme_def}{definition of the custom scheme}

  \item{k}{peptide lengths to include}

  \item{n}{exponents of the library size to include}
}
\value{
A data frame of library information
}
\description{
For a given scheme, generate a dataset with the library
information
}

