\name{generateCustomProbs}
\alias{generateCustomProbs}
\title{For a given scheme, generate a dataset with the peptide probabilities}
\usage{
  generateCustomProbs(scheme_def)
}
\arguments{
  \item{scheme_def}{definition of the custom scheme}
}
\value{
  A data frame of peptide probabilities
}
\description{
  For a given scheme, generate a dataset with the peptide
  probabilities
}

