% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generateCustomProbs}
\alias{generateCustomProbs}
\title{For a given scheme, generate a dataset with the peptide probabilities}
\usage{
generateCustomProbs(scheme_def, k = 6:10)
}
\arguments{
\item{scheme_def}{definition of the custom scheme}

\item{k}{peptide lengths to include}
}
\value{
A data frame of peptide probabilities
}
\description{
For a given scheme, generate a dataset with the peptide probabilities
}

