% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tabulate.r
\name{generateCustomProbs}
\alias{generateCustomProbs}
\title{For a given scheme, generate a dataset with the peptide probabilities}
\usage{
generateCustomProbs(scheme_def, k = 1:20)
}
\arguments{
\item{scheme_def}{definition of the custom scheme}

\item{k}{peptide lengths to include}
}
\value{
A data frame of peptide probabilities
}
\description{
For a given scheme, generate a dataset with the peptide probabilities
}

