% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tabulate.r
\name{getChoices}
\alias{getChoices}
\title{Get the number of peptides that reduce to a particular reduced encoding}
\usage{
getChoices(str)
}
\arguments{
\item{str}{The reduced encoding string}
}
\value{
An integer of the possible number of peptides reducing to this encoding
}
\description{
Get the number of peptides that reduce to a particular reduced encoding
}

