\docType{data}
\name{nnn_scheme}
\alias{nnn_scheme}
\title{Nucleotide library scheme NNN}
\format{'data.frame':	6 obs. of  5 variables:
 $ class : Factor w/ 6 levels "A","B","C","D",..: 1 2 3 4 5 6
 $ aacids: Factor w/ 6 levels "AGPTV","C*","DEFHKNQY",..: 6 1 4 3 5 2
 $ s     : num  3 5 1 8 2 2
 $ c     : num  6 4 3 2 1 2.5
 $ stops : num  3 3 3 3 3 3}
\description{
  This data set contains descriptions of amino acid classes
  under the NNN library scheme.
}
\details{
  Nucleotide library scheme NNN
}
\keyword{datasets}

